requirements.txt
fastapi
uvicorn
sqlalchemy
databases
psycopg2-binary
pydantic
alembic
python-dotenv